fix: move configcheck to service package
